Change key equivalent for inserting a newline comment to ^⌘⏎
Using  ⌘⏎ conflicts with the built-in TextMate macro to "Move to the EOL and Insert LF". While this conflict only occurs within line-comments (due to the scope set by the snippet), it is easily resolved by adding the "control" modifier character, preserving the conventional behavior while offering the additional functionality to those who wish to have it.